% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.default.R, R/assess.highReg.R,
%   R/assess.lowClass.R, R/assess.lowReg.R
\name{assess.default}
\alias{assess.default}
\alias{assess}
\alias{assess.highReg}
\alias{assess.lowClass}
\alias{assess.lowReg}
\title{Create a model fit based on the data.}
\usage{
\method{assess}{default}(x, ...)

assess(x, ...)

\method{assess}{highReg}(x, ...)

\method{assess}{lowClass}(x, ...)

\method{assess}{lowReg}(x, ...)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}

\item{...}{Additional arguments}
}
\value{
model (object) The model of a data
}
\description{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
