% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_data.R
\docType{data}
\name{chem_data}
\alias{chem_data}
\title{Chemical data obtained from the UCI Machine Learning data base}
\format{
## `chem_data`
A data frame with 178 rows and 14 columns:
}
\source{
<https://archive.ics.uci.edu/ml/datasets/Wine>
}
\usage{
chem_data
}
\description{
Wine data set.
}
\keyword{datasets}
