% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classLOGISTIC.R
\name{classLOGISTIC}
\alias{classLOGISTIC}
\title{Linear regression}
\usage{
classLOGISTIC(x)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Wrapper for linear Regression
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
