% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMLP.R
\name{classMLP}
\alias{classMLP}
\title{MLP sklearn classifier}
\usage{
classMLP(x)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Wrapper for scikit-learns MLP classifier.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
