% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRF.R
\name{classRF}
\alias{classRF}
\title{Classification using Random Forest}
\usage{
classRF(x)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Classification using Random Forest
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
