% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSVM.R
\name{classSVM}
\alias{classSVM}
\title{Classification using Support Vector Machine}
\usage{
classSVM(x)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Classification using Support Vector Machine
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
