% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.confimClass.R, R/confirm.confirmReg.R,
%   R/confirm.default.R
\name{confirm.confirmClass}
\alias{confirm.confirmClass}
\alias{confirm.confirmReg}
\alias{confirm}
\title{Create a set of prediction summary values based on the class.}
\usage{
\method{confirm}{confirmClass}(x, ...)

\method{confirm}{confirmReg}(x, ...)

confirm(x, ...)
}
\arguments{
\item{x}{(list of vectors) A list containing the observed and predicted
values.}

\item{...}{Additional parameters}
}
\value{
predict_summary (list) The summary prediction values.
}
\description{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
