% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{Confusion Matrix}
\usage{
confusion(observed, predicted)
}
\arguments{
\item{observed}{(binary vector) The observed values for the response.}

\item{predicted}{(binary vector) The fitted values developed by the model.}
}
\value{
confusion (integer data frame) The confusion matrix per class.
}
\description{
Build the confusion matrix. Predicted as columns and observed as rows.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
