% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denormalize.R
\name{denormalize}
\alias{denormalize}
\title{Denormalize}
\usage{
denormalize(x, x_mean, x_sd)
}
\arguments{
\item{x}{(matrix, data frame, or scalar vector) The numeric values to be scaled.}

\item{x_mean}{(scalar vector or float) The column means of x.}

\item{x_sd}{(scalar vector or float) The column standard deviations of x.}
}
\value{
x (matrix, data frame, or scalar vector) The descaled values of x.
}
\description{
Multiply by the standard deviation and add the mean.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
