% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreDetectAnomaly.R
\name{exploreDetectAnomaly}
\alias{exploreDetectAnomaly}
\title{Detect time series outliers}
\usage{
exploreDetectAnomaly(x)
}
\arguments{
\item{x}{(vector) A vector of values in which there are suspected outliers.}
}
\value{
A vector of indicies of suspected outliers.
}
\description{
Detect time series outliers. Each column is assumed to be a time series.
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
