% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreImpute.R
\name{exploreImpute}
\alias{exploreImpute}
\title{Multiple imputation of missing values}
\usage{
exploreImpute(x, seedval = NULL)
}
\arguments{
\item{x}{(matrix) A matrix of values in which there are NA values.}

\item{seedval}{(integer) The seed for imputation.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Multiple imputation of missing values.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
