% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreImputeRF.R
\name{exploreImputeRF}
\alias{exploreImputeRF}
\title{Multiple imputation of missing values}
\usage{
exploreImputeRF(x, y, ntree = 500)
}
\arguments{
\item{x}{(matrix) A matrix of values in which there are NA values.}

\item{y}{(vector) A response to predict based on the independent variables.}

\item{ntree}{(integer) The number of trees in the forest.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Multiple imputation of missing values.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
