% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreTSoutliers.R
\name{exploreTSoutliers}
\alias{exploreTSoutliers}
\title{Remove time series outliers}
\usage{
exploreTSoutliers(x)
}
\arguments{
\item{x}{(matrix) A matrix of values in which there are suspected outliers.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Remove time series outliers. Each column is assumed to be a time series.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
