% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holistic_model.R, R/regENET.R, R/regGAM.R,
%   R/regGLM.R, R/regLM.R, R/regMLP.R, R/regPCR.R, R/regRF.R, R/regSVM.R
\name{holistic_model.default}
\alias{holistic_model.default}
\alias{holistic_model}
\alias{holistic_model.regGLMNET}
\alias{holistic_model.regGAM}
\alias{holistic_model.regGLM}
\alias{holistic_model.regLM}
\alias{holistic_model.regMLP}
\alias{holistic_model.regPCR}
\alias{holistic_model.regRF}
\alias{holistic_model.regSVM}
\title{Create holistic documentation for a given model.}
\usage{
\method{holistic_model}{default}(object, ...)

holistic_model(object, ...)

\method{holistic_model}{regGLMNET}(object, ...)

\method{holistic_model}{regGAM}(object, ...)

\method{holistic_model}{regGLM}(object, ...)

\method{holistic_model}{regLM}(object, ...)

\method{holistic_model}{regMLP}(object, ...)

\method{holistic_model}{regPCR}(object, ...)

\method{holistic_model}{regRF}(object, ...)

\method{holistic_model}{regSVM}(object, ...)
}
\arguments{
\item{object}{(object) A fit object.}

\item{...}{Additional arguments}
}
\value{
doc (string) The documentation for a given model.
}
\description{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
