% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holistic_models.R
\name{holistic_models}
\alias{holistic_models}
\title{Holistic model assessment}
\usage{
holistic_models(x, y)
}
\arguments{
\item{x}{(data frame or matrix) A set of independent variables.}

\item{y}{(vector) The response.}
}
\value{
metrics (list of metrics per model) The metrics associated with each
metric and model.
}
\description{
Wrapper for comparing all models.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
