% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holistic_splits.R
\name{holistic_splits}
\alias{holistic_splits}
\title{Holistic model assessment with multiple splits}
\usage{
holistic_splits(x, y, num_splits = 5, n_core = NULL)
}
\arguments{
\item{x}{(data frame or matrix) A set of independent variables.}

\item{y}{(vector) The response.}

\item{num_splits}{(integer) The total number of iterations of splits.}

\item{n_core}{(integer) The total number of cores to use.}
}
\value{
metrics_list (list of metrics per model) The metrics associated with
each metric and model and split.
}
\description{
Wrapper for comparing all models over each split.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
