% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_model.R
\name{identify_model}
\alias{identify_model}
\title{Identify models}
\usage{
identify_model(x, y)
}
\arguments{
\item{x}{(data frame or matrix) A set of independent variables.}

\item{y}{(vector) The response.}
}
\value{
models (list of string vectors) The models and metrics to run.
}
\description{
Identify a list of models and metrics based on the data set.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
