% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.high_dimensional.R
\name{is.high_dimensional}
\alias{is.high_dimensional}
\title{Is the data high dimensional}
\usage{
is.high_dimensional(x)
}
\arguments{
\item{x}{(data frame or matrix) A set of independent variables.}
}
\value{
is.low (bool) Is x low dimensional.
}
\description{
Utils function to check if high-dimensional $(n < p)$, i.e., the number of
rows are far less than the number of parameters.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
