% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_summary.R
\name{list_to_summary}
\alias{list_to_summary}
\title{Create a list of summary tables from list.}
\usage{
list_to_summary(metrics_list, digits = 3)
}
\arguments{
\item{metrics_list}{(list of metrics per model) The metrics associated with
each metric and model and split.}

\item{digits}{(string) The number of digits to round to for each metric.}
}
\value{
df_list (list of data frames) The list of summary data frames.
}
\description{
Break down a set of summary data frames by each metric, model, and set.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
