% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_to_df.R
\name{lists_to_df}
\alias{lists_to_df}
\title{Create a list of data frames from list.}
\usage{
lists_to_df(metrics_list, column = "metric")
}
\arguments{
\item{metrics_list}{(list of metrics per model) The metrics associated with
each metric and model and split.}

\item{column}{(string) The column indicator to create each list element.}
}
\value{
df_list (list of data frames) The list of data frames for each
unique element in the column argument.
}
\description{
Break down a set of data frames by either a metric, model, or set.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
