% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape}
\alias{mape}
\title{mape}
\usage{
mape(observed, predicted)
}
\arguments{
\item{observed}{(scalar vector) The observed values for the response.}

\item{predicted}{(scalar vector) The fitted values developed by the model.}
}
\value{
mape (scalar float) The mape between the observed and predicted.
}
\description{
Mean absolute percent error.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
