% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x, params = NULL, norm_params = NULL)
}
\arguments{
\item{x}{(matrix, data frame, or scalar vector) The numeric values to be
normalized.}

\item{params}{(list) The associated values for the normalization process.}

\item{norm_params}{(list) The associated values required for reversing the
normalization process.}
}
\value{
result (list) The normalized values and normalization parameters.
}
\description{
Subtract the mean and divide by the standard deviation. If given [params],
then will apply a given mean and standard deviation to the data set.
Otherwise, will calculate the mean and standard deviation. If [norm_params]
is not NULL, will reverse the normalization process.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
