% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeBinomial.R
\name{normalizeBinomial}
\alias{normalizeBinomial}
\title{Normalize Binomial}
\usage{
normalizeBinomial(x, params = NULL, norm_params = NULL)
}
\arguments{
\item{x}{(matrix, data frame, or scalar vector) The numeric values to be
normalized.}

\item{params}{(list) The associated values for the normalization process.}

\item{norm_params}{(list) The associated values required for reversing the
normalization process.}
}
\value{
result (list) The normalized values and normalization parameters.
}
\description{
Transform the variable to be between zero and one. If [norm_params] is NULL, then
will normalize the data. Otherwise, will reverse the normalization process.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
