% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin_metric.R
\name{plot_violin_metric}
\alias{plot_violin_metric}
\title{Plot violin metric}
\usage{
plot_violin_metric(df)
}
\arguments{
\item{df}{(data frame) The numeric values to be scaled.}
}
\value{
fig (plotly figure) The descaled values of x.
}
\description{
Plot violin metrics for a given data frame with the following column names:
model, value, and set. Model indicates the machine learning model used,
value is the corresponding value of the metric used (e.g., rmse), and set
determines if the metric was from the training or test set.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
