% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.regGAM.R
\name{predict.regGAM}
\alias{predict.regGAM}
\title{Predict Generalized Linear Model}
\usage{
\method{predict}{regGAM}(object, x, ...)
}
\arguments{
\item{object}{(object) A fit object.}

\item{x}{(list) A list containing [x], the set of independent variables and
[y], the response.}

\item{...}{Additional arguments.}
}
\value{
predicted (list of scalar vectors) The predicted values of train and
test.
}
\description{
Predict Generalized Linear Model.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
\seealso{
predict.gam
}
