% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.regGLMNET.R
\name{predict.regGLMNET}
\alias{predict.regGLMNET}
\title{Predict function for all glmnet models}
\usage{
\method{predict}{regGLMNET}(object, x, ...)
}
\arguments{
\item{object}{(object) A fit object.}

\item{x}{(list) A list containing [x], the set of independent variables and
[y], the response.}

\item{...}{Additional arguments.}
}
\value{
predicted (list of scalar vectors) The predicted values of train and
test.
}
\description{
Prediction for glmnet regression models including: enet, lasso,
and ridge.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
\seealso{
predict.glmnet
}
