% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.regLM.R
\name{predict.regLM}
\alias{predict.regLM}
\title{Linear regression}
\usage{
\method{predict}{regLM}(object, x, ...)
}
\arguments{
\item{object}{(object) A fit object.}

\item{x}{(list) A list containing [x], the set of independent variables and
[y], the response.}

\item{...}{Additional arguments.}
}
\value{
predicted (list of scalar vectors) The observed and predicted values.
}
\description{
Wrapper for linear regression.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
\seealso{
predict.lm
}
