% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.try_error.R
\name{predict.try_error}
\alias{predict.try_error}
\title{Predict try error}
\usage{
\method{predict}{try_error}(object, x, ...)
}
\arguments{
\item{object}{(object) A fit object.}

\item{x}{(list) A list containing [x], the set of independent variables and
[y], the response.}

\item{...}{Additional arguments.}
}
\value{
predicted (list of scalar vectors) The predicted values of train and
test.
}
\description{
Returns a set of zeros.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
