% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictInit.R
\name{predictInit}
\alias{predictInit}
\title{Initialize Prediction}
\usage{
predictInit(train, test = NULL)
}
\arguments{
\item{train}{(data frame or matrix) A set of independent variables.}

\item{test}{(data frame or matrix) A set of independent variables for
test set.}
}
\value{
predicted (list of scalar vectors) The model predicted.
}
\description{
Create an initial set of zeros for the train and test prediction.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
