% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{r2}
\usage{
r2(observed, predicted)
}
\arguments{
\item{observed}{(scalar vector) The observed values for the response.}

\item{predicted}{(scalar vector) The fitted values developed by the model.}
}
\value{
r2 (scalar float) The r2 between the observed and predicted.
}
\description{
Coefficient of determination
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
