% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDMD.R
\name{reduceDMD}
\alias{reduceDMD}
\title{Reduce via the Dynamic Mode Decomposition}
\usage{
reduceDMD(x, rank = NULL)
}
\arguments{
\item{x}{(matrix) The matrix that will be reduced.}

\item{rank}{(interger | NULL) The estimated rank of the matrix. Will be
determined using Donoho's threshold if left NULL.}
}
\value{
x_reduced (matrix) The dimension reduced matrix.
}
\description{
Reduce via the Dynamic Mode Decomposition
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
