% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceGAM.R
\name{reduceGAM}
\alias{reduceGAM}
\title{Reduce the via STL}
\usage{
reduceGAM(x, n, nknots = NULL)
}
\arguments{
\item{x}{(matrix) The matrix that will be reduced.}

\item{n}{(integer) The desired number of rows.}

\item{nknots}{(integer) The number of knots to use for the spline fit.}
}
\value{
x_reduced (matrix) The dimension reduced matrix.
}
\description{
Reduce the via STL
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
