% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceINTERP.R
\name{reduceINTERP}
\alias{reduceINTERP}
\title{Reduce the via column-wise interpolation}
\usage{
reduceINTERP(x, n)
}
\arguments{
\item{x}{(matrix) The matrix that will be reduced.}

\item{n}{(integer) The desired number of rows.}
}
\value{
x_reduced (matrix) The dimension reduced matrix.
}
\description{
Reduce the via column-wise interpolation
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
