% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceSVD.R
\name{reduceSVD}
\alias{reduceSVD}
\title{Reduce via the Singular Value Decomposition}
\usage{
reduceSVD(x, rank = NULL, scaledata = FALSE)
}
\arguments{
\item{x}{(matrix) The matrix that will be reduced.}

\item{rank}{(interger | NULL) The estimated rank of the matrix. Will be
determined using Donoho's threshold if left NULL.}

\item{scaledata}{(boolean) Indicator to scale the data prior to the SVD.}
}
\value{
x_reduced (matrix) The dimension reduced matrix.
}
\description{
Reduce via the Singular Value Decomposition
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
