% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceVIF.R
\name{reduceVIF}
\alias{reduceVIF}
\title{Reduce via Variance Inflation Factor}
\usage{
reduceVIF(x, threshold = 0.95)
}
\arguments{
\item{x}{(matrix) The matrix that will be reduced.}

\item{threshold}{(numeric scalar between 0 and 1) The maximum correlation for
a given variable.}
}
\value{
x_reduced (matrix) The dimension reduced matrix.
}
\description{
Reduce via Variance Inflation Factor
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
