% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regENET.R
\name{regENET}
\alias{regENET}
\title{regGLMNET}
\usage{
regENET(
  x,
  lambda = NULL,
  alpha = 0.5,
  family = c("gaussian", "binomial", "poisson", "mguassian")
)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}

\item{lambda}{(numeric vector) The weight of penalization on the regression
coefficients. Can be NULL where lambda is determined through theoretical
estimates or be provided as a vector of values to optimize over.}

\item{alpha}{(numeric between 0 and 1) The alpha parameter controls the
type of penalization. A value of 0 indicates ridge regression (only dealing
with multicollinearity), a value of 1 indicates lasso regression (only
providing variable selection), and a value of 0.5 indicates elastic net
(multicollinearity and variable selection)."}

\item{family}{(string) The distribution associated with the errors, e.g.,
gaussian for a linear model or binomial for logistic.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via Generalized Linear Models and Elastic Net.
}
\details{
This is an extension of a linear model through the combination of a
generalized linear model and penalization of the regression coefficients.
Note that an alpha value of 0 indicates ridge regression while an alpha value
of 1 indicates lasso regression.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
