% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regGAM.R
\name{regGAM}
\alias{regGAM}
\title{regGAM}
\usage{
regGAM(
  x,
  k = 10,
  family = c("gaussian", "binomial", "quasibinomial", "poisson", "quasipoisson")
)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}

\item{k}{(integer) Number of knots associated with each variable when
creating a smooth function.}

\item{family}{(string) The distribution associated with the errors, e.g.,
"gaussian" for a linear model or "binomial" for logistic.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via Generalized Additive Models
}
\details{
The Generalized Additive Model (GAM) is an extension of the generalized
linear model while assuming the independent variables can be represented
as smooth functions.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
