% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regGLM.R
\name{regGLM}
\alias{regGLM}
\title{regGLM}
\usage{
regGLM(
  x,
  family = c("gaussian", "binomial", "quasibinomial", "poisson", "quasipoisson")
)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}

\item{family}{(string) The distribution associated with the errors, e.g.,
"gaussian" for a linear model or "binomial" for logistic.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via Generalized Linear Models
}
\details{
The Generalized Linear Model (GLM) is an extension of a linear model where the
independent variables and the dependent variable share a potentially
non-Gaussian relationship.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
