% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regLM.R
\name{regLM}
\alias{regLM}
\title{regLM}
\usage{
regLM(x)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via a Linear Model
}
\details{
Application of a Ordinary Least Squares model.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
