% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regMLP.R
\name{regMLP}
\alias{regMLP}
\title{MLP sklearn}
\usage{
regMLP(x, activation = c("relu", "identity", "logistic", "tanh"))
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}

\item{activation}{(string) The activation function between nodes.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via Multi-Layer Perceptron
}
\details{
The Multi-Layer Perceptron (MLP) is a feed forward artificial neural network.
This function utilizes the scikit-learn MLPRegressor function.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
