% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPCR.R
\name{regPCR}
\alias{regPCR}
\title{regPCR}
\usage{
regPCR(x)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via a Principal Component Regression
}
\details{
Application of dimension reduction and regression.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
