% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regRF.R
\name{regRF}
\alias{regRF}
\title{regRF}
\usage{
regRF(x, ntree = 500)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}

\item{ntree}{(integer) The total number of trees to use in building the
random forest.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via Random Forest
}
\details{
The Random Forest (RF) technique is an ensemble learning method by creating
multiple decision trees.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
