% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regSVM.R
\name{regSVM}
\alias{regSVM}
\title{regSVM}
\usage{
regSVM(
  x,
  kernel = c("linear", "polynomial", "radial", "sigmoid"),
  gamma = NULL
)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables as a
data frame and [y], the response as a vector.}

\item{kernel}{(string) The kernel associated with the nonlinear behaviors of
the data set.}

\item{gamma}{(numeric scalar) A tuning parameter based on the kernel. The
linear kernel does not require a value of gamma.}
}
\value{
fit (model specific object) The model fit.
}
\description{
Regression via Support Vector Machine
}
\details{
The Support-Vector Machine (SVM) technique is based on optimizing a specific
hyperplane (or sets of hyperplanes) between the data set variables.

Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
