% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(observed, predicted)
}
\arguments{
\item{observed}{(scalar vector) The observed values for the response.}

\item{predicted}{(scalar vector) The fitted values developed by the model.}
}
\value{
rmse
}
\description{
Root mean square error.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
