% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDataClass.R
\name{setDataClass}
\alias{setDataClass}
\title{Set Data Class}
\usage{
setDataClass(x)
}
\arguments{
\item{x}{(list) A list containing [x], the set of independent variables and
[y], the response.}
}
\value{
result (list) The training and test as a data frame.
}
\description{
Verify the dimensions of the data as well as determines the appropriate
class for the data.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
