% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smape.R
\name{smape}
\alias{smape}
\title{smape}
\usage{
smape(observed, predicted)
}
\arguments{
\item{observed}{(scalar vector) The observed values for the response.}

\item{predicted}{(scalar vector) The fitted values developed by the model.}
}
\value{
smape
}
\description{
Symmetrical mean absolute percent error
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
