% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdThreshold.R
\name{svdThreshold}
\alias{svdThreshold}
\title{SVD threshold}
\usage{
svdThreshold(singularvalues, p = NULL)
}
\arguments{
\item{singularvalues}{The singular values obtained through the SVD.}

\item{p}{The number of columns of the matrix.}
}
\value{
rank (integer) The approximate rank of the matrix.
}
\description{
Determine the appropriate threshold for the singular values.
Donoho, David L., and Matan Gavish. "The optimal hard threshold for singular values is 4/sqrt(3)." arXiv preprint arXiv:1305.5870 4 (2013).
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
