% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classification_data_no_test.R
\name{test_classification_data_no_test}
\alias{test_classification_data_no_test}
\title{Classification test data}
\usage{
test_classification_data_no_test()
}
\value{
data (list) The model independent variables and response.
}
\description{
Generate the classification test data from chem_data.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
