% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_regression_data_no_test.R
\name{test_regression_data_no_test}
\alias{test_regression_data_no_test}
\title{Regression test data}
\usage{
test_regression_data_no_test()
}
\value{
data (list) The model independent variables and response.
}
\description{
Generate the regression test data from chem_data.
}
\details{
Copyright 2023 Battelle Energy Alliance ALL RIGHTS RESERVED
}
